# test: Shorten genai-perf test time, fail fast on server startup, and upgrade to 24.06