--- a/.pre-commit-config.yaml
+++ b/.pre-commit-config.yaml
@@ -27,12 +27,18 @@
 repos:
 
 # Python formatting/linting
-## Run the Ruff linter and formatter.
+## Run the Ruff formatter.
+- repo: https://github.com/astral-sh/ruff-pre-commit
+  rev: v0.5.2
+  hooks:
+    - id: ruff-format
+      exclude: "src/triton_cli/trt_llm/checkpoint_scripts|src/triton_cli/templates/trt_llm"
+
+## Run the Ruff linter.
 - repo: https://github.com/astral-sh/ruff-pre-commit
   rev: v0.5.2
   hooks:
     - id: ruff
-    - id: ruff-format
       exclude: "src/triton_cli/trt_llm/checkpoint_scripts|src/triton_cli/templates/trt_llm"
 
 # Upgrade Python syntax