This PR makes the following changes:
1. Adds `--request-count 10` to `genai-perf` when running `triton profile` in the test scripts. This is done to reduce risk of github action jobs timing out, use less overall github action minutes, and speed up local development time when running tests locally.

2. Removes the `triton profile` call for the `http` subtest, since `genai-perf` is using `grpc` under the hood for both tests, it was just wasting test time without adding value.

3. Improves the checks around server readiness to detect if the server process failed to start altogether, rather than waiting the full timeout period only polling the endpoints.

Note that locally, `--request-count 1` and `--request-count 10` both took about the same time of ~4-5 seconds instead of the previous default behavior taking 1m30s - 2m or so.

Overall, this should reduce the e2e test time by about 3-5 minutes in the succesful case, and possibly even more when server startups fail for some reason.

---

Example successful tests with genai-perf speedup changes, full test time is ~50seconds:

```
root@ced35d0-lcedt:/mnt/triton/cli/triton_cli/tests# pytest -v
===================================================================================== test session starts =====================================================================================
platform linux -- Python 3.10.12, pytest-8.2.2, pluggy-1.5.0 -- /usr/bin/python3
cachedir: .pytest_cache
rootdir: /mnt/triton/cli/triton_cli
configfile: pyproject.toml
plugins: timeout-2.3.1, mock-3.14.0
collected 55 items

test_cli.py::TestRepo::test_clear[None] PASSED                                                                                                                                          [  1%]
test_cli.py::TestRepo::test_clear[tmp/models] PASSED                                                                                                                                    [  3%]
test_cli.py::TestRepo::test_import_known_model[None-llama-2-7b] PASSED                                                                                                                  
...
test_cli.py::TestRepo::test_import_vllm[vllm-model-hf:gpt2] PASSED                                                                                                                      [ 63%]
test_cli.py::TestRepo::test_repo_add_trtllm_build[trtllm-model-hf:gpt2] SKIPPED (Only run for TRT-LLM image)                                                                            [ 65%]
test_cli.py::TestRepo::test_import_trtllm_prebuilt SKIPPED (Pre-built TRT-LLM engines not available)                                                                                    [ 67%]
test_cli.py::TestRepo::test_import_no_source PASSED                                                                                                                                     [ 69%]
test_cli.py::TestRepo::test_remove PASSED                                                                                                                                               [ 70%]
test_cli.py::TestRepo::test_remove_nonexistent PASSED                                                                                                                                   [ 72%]
test_cli.py::TestRepo::test_list[None] PASSED                                                                                                                                           [ 74%]
test_cli.py::TestRepo::test_list[tmp/models] PASSED                                                                                                                                     [ 76%]
test_cli.py::TestRepo::test_triton_profile PASSED                                                                                                                                       [ 78%]
test_cli.py::TestRepo::test_triton_metrics[mock_llm] PASSED                                                                                                                             [ 80%]
test_cli.py::TestRepo::test_triton_config[mock_llm] PASSED                                                                                                                              [ 81%]
test_cli.py::TestRepo::test_triton_status[mock_llm] PASSED                                                                                                                              [ 83%]
test_e2e.py::TestE2E::test_tensorrtllm_e2e[grpc] SKIPPED (Only run for TRT-LLM image)                                                                                                   [ 85%]
test_e2e.py::TestE2E::test_tensorrtllm_e2e[http] SKIPPED (Only run for TRT-LLM image)                                                                                                   [ 87%]
test_e2e.py::TestE2E::test_vllm_e2e[grpc] SKIPPED (Only run for VLLM image)                                                                                                             [ 89%]
test_e2e.py::TestE2E::test_vllm_e2e[http] SKIPPED (Only run for VLLM image)                                                                                                             [ 90%]
test_e2e.py::TestE2E::test_non_llm[grpc] PASSED                                                                                                                                         [ 92%]
test_e2e.py::TestE2E::test_non_llm[http] PASSED                                                                                                                                         [ 94%]
test_e2e.py::TestE2E::test_mock_llm[grpc] PASSED                                                                                                                                        [ 96%]
test_e2e.py::TestE2E::test_mock_llm[http] PASSED                                                                                                                                        [ 98%]
test_library.py::test_version PASSED                                                                                                                                                    [100%]

=============================================================================== 49 passed, 6 skipped in 50.46s ================================================================================
```

Example run with server startup failures, failing fast rather than waiting 60 seconds on each failed startup polling the endpoints:
```
================================================================================== short test summary info ===================================================================================
FAILED test_cli.py::TestRepo::test_triton_metrics[mock_llm] - Exception: 'triton start' PID 824213 was in a zombie state.
FAILED test_cli.py::TestRepo::test_triton_config[mock_llm] - Exception: 'triton start' PID 824227 was in a zombie state.
FAILED test_cli.py::TestRepo::test_triton_status[mock_llm] - Exception: 'triton start' PID 824242 was in a zombie state.
FAILED test_e2e.py::TestE2E::test_mock_llm[grpc] - Exception: 'triton start' PID 824254 was in a zombie state.
FAILED test_e2e.py::TestE2E::test_mock_llm[http] - Exception: 'triton start' PID 824266 was in a zombie state.
======================================================================== 5 failed, 50 deselected, 1 warning in 5.77s =========================================================================
```